+++ plugins/multichain/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @zerodev/multi-chain-sdk
 
+## 5.3.2
+
+### Patch Changes
+
+- Removed redundant calls to passkey-server during message signing
+
+## 5.3.1
+
+### Patch Changes
+
+- Update viem to 2.16.3
+
 ## 5.3.0
 
 ### Minor Changes